update vite/vue stack
vite-plugin-vue2 is deprecated, swap to @vitejs/plugin-vue2

solves the errors:
Sourcemap is likely to be incorrect: a plugin (vite:worker) was used to transform files, but didn't generate a sourcemap for the transformation. Consult the plugin documentation for help
Sourcemap is likely to be incorrect: a plugin (vite-plugin-md) was used to transform files, but didn't generate a sourcemap for the transformation. Consult the plugin documentation for help